ice: add flag to distinguish reset from .ndo_bpf in XDP rings config
Add a bool argument to both ice_prepare_xdp_rings() and
ice_destroy_xdp_rings() in order to distinguish between calls from rebuild
and .ndo_bpf().

The behaviour of the XDP rings config functions is already
context-dependent, but the logic is to guess it, this leads to some
synchronization problems. Particularly, when:

* PFR is requested by tx_timeout handler
* .ndo_bpf() is asked to delete the program, calls ice_destroy_xdp_rings(),
  but reset flag is set, so rings are destroyed without deleting the
  program
* ice_vsi_rebuild tries to delete non-existent XDP rings, because the
  program is still on the VSI
* system crashes

Fixes: efc2214b6047 ("ice: Add support for XDP")
Signed-off-by: Larysa Zaremba <larysa.zaremba@intel.com>